+++ src/trajectory_initialization.jl
@@ -3,8 +3,7 @@ module TrajectoryInitialization
 export unitary_geodesic
 export linear_interpolation
 export unitary_linear_interpolation
-export initialize_unitary_trajectory
-export initialize_quantum_state_trajectory
+export initialize_trajectory
 export convert_fixed_time
 export convert_free_time
 
@@ -118,31 +117,40 @@ function unitary_geodesic(
     if U_goal isa EmbeddedOperator
         U_goal = U_goal.operator
     end
-    return unitary_geodesic(Matrix{ComplexF64}(I(size(U_goal, 1))), U_goal, samples; kwargs...)
+    return unitary_geodesic(
+        Matrix{ComplexF64}(I(size(U_goal, 1))),
+        U_goal,
+        samples;
+        kwargs...
+    )
 end
 
+"""
+    unitary_geodesic(U_init, U_goal, times; kwargs...)
+
+Compute the geodesic connecting U_init and U_goal at the specified times. Allows for the possibility of unequal times and ranges outside [0,1].
+
+# Arguments
+- `U_init::AbstractMatrix{<:Number}`: The initial unitary operator.
+- `U_goal::AbstractMatrix{<:Number}`: The goal unitary operator.
+- `times::AbstractVector{<:Number}`: The times at which to evaluate the geodesic.
+
+# Keyword Arguments
+- `return_unitary_isos::Bool=true`: If true returns a matrix where each column is a unitary isovec, i.e. vec(vcat(real(U), imag(U))). If false, returns a vector of unitary matrices.
+- `return_generator::Bool=false`: If true, returns the effective Hamiltonian generating the geodesic.
+"""
 function unitary_geodesic(
     U_init::AbstractMatrix{<:Number},
     U_goal::AbstractMatrix{<:Number},
-    timesteps::AbstractVector{<:Number};
+    times::AbstractVector{<:Number};
     return_unitary_isos=true,
     return_generator=false
 )
-    """
-    Compute the effective generator of the geodesic connecting U₀ and U₁.
-        U_goal = exp(-im * H * T) U_init
-        log(U_goal * U_init') = -im * H * T
-
-    Allow for the possibiltiy of unequal timesteps and ranges outside [0,1].
-
-    Returns the geodesic.
-    Optionally returns the effective Hamiltonian generating the geodesic.
-    """
-    t₀ = timesteps[1]
-    T = timesteps[end] - t₀
+    t₀ = times[1]
+    T = times[end] - t₀
     H = im * log(U_goal * U_init') / T
     # -im prefactor is not included in H
-    U_geo = [exp(-im * H * (t - t₀)) * U_init for t ∈ timesteps]
+    U_geo = [exp(-im * H * (t - t₀)) * U_init for t ∈ times]
     if !return_unitary_isos
         if return_generator
             return U_geo, H
@@ -167,7 +175,7 @@ linear_interpolation(x::AbstractVector, y::AbstractVector, n::Int) =
 const VectorBound = Union{AbstractVector{R}, Tuple{AbstractVector{R}, AbstractVector{R}}} where R <: Real
 const ScalarBound = Union{R, Tuple{R, R}} where R <: Real
 
-function initialize_unitaries(
+function initialize_unitary_trajectory(
     U_init::AbstractMatrix{<:Number},
     U_goal::OperatorType,
     T::Int;
@@ -185,7 +193,7 @@ end
 #                           Initial controls                                    #
 # ----------------------------------------------------------------------------- #
 
-function initialize_controls(
+function initialize_control_trajectory(
     n_drives::Int,
     n_derivatives::Int,
     T::Int,
@@ -216,22 +224,28 @@ function initialize_controls(
     return controls
 end
 
-function initialize_controls(a::AbstractMatrix, Δt::AbstractVecOrMat, n_derivatives::Int)
+function initialize_control_trajectory(
+    a::AbstractMatrix,
+    Δt::AbstractVecOrMat,
+    n_derivatives::Int
+)
     controls = Matrix{Float64}[a]
+
     for n in 1:n_derivatives
         # next derivative
         push!(controls,  derivative(controls[end], Δt))
 
         # to avoid constraint violation error at initial iteration for da, dda, ...
         if n > 1
-            controls[end-1][:, end] = controls[end-1][:, end-1] + Δt[end-1] * controls[end][:, end-1]
+            controls[end-1][:, end] =
+                controls[end-1][:, end-1] + Δt[end-1] * controls[end][:, end-1]
         end
     end
     return controls
 end
 
-initialize_controls(a::AbstractMatrix, Δt::Real, n_derivatives::Int) =
-    initialize_controls(a, fill(Δt, size(a, 2)), n_derivatives)
+initialize_control_trajectory(a::AbstractMatrix, Δt::Real, n_derivatives::Int) =
+    initialize_control_trajectory(a, fill(Δt, size(a, 2)), n_derivatives)
 
 # ----------------------------------------------------------------------------- #
 #                           Trajectory initialization                           #
@@ -245,28 +259,45 @@ Initialize a trajectory for a unitary control problem. The trajectory is initial
 data that should be consistently the same type (in this case, Float64).
 
 """
-function initialize_unitary_trajectory(
-    U_goal::OperatorType,
+function initialize_trajectory(
+    state_data::Vector{Matrix{Float64}},
+    state_inits::Vector{<:AbstractVector{Float64}},
+    state_goals::Vector{<:AbstractVector{Float64}},
+    state_names::AbstractVector{Symbol},
     T::Int,
     Δt::Union{Float64, AbstractVecOrMat{<:Float64}},
     n_drives::Int,
-    all_a_bounds::NamedTuple{anames, <:Tuple{Vararg{VectorBound}}} where anames;
-    U_init::AbstractMatrix{<:Number}=Matrix{ComplexF64}(I(size(U_goal, 1))),
-    n_derivatives::Int=2,
-    geodesic=true,
-    bound_unitary=false,
+    control_bounds::Tuple{Vararg{VectorBound}};
+    bound_state=false,
     free_time=false,
+    control_name=:a,
+    n_control_derivatives::Int=length(control_bounds) - 1,
+    timestep_name=:Δt,
     Δt_bounds::ScalarBound=(0.5 * Δt, 1.5 * Δt),
     drive_derivative_σ::Float64=0.1,
     a_guess::Union{AbstractMatrix{<:Float64}, Nothing}=nothing,
-    system::Union{AbstractQuantumSystem, AbstractVector{<:AbstractQuantumSystem}, Nothing}=nothing,
     global_data::Union{NamedTuple, Nothing}=nothing,
-    rollout_integrator::Function=expv,
-    Ũ⃗_keys::AbstractVector{<:Symbol}=[:Ũ⃗],
-    a_keys::AbstractVector{<:Symbol}=[Symbol("d"^i * "a") for i in 0:n_derivatives]
 )
-    @assert length(a_keys) == n_derivatives + 1 "a_keys must have the same length as n_derivatives + 1"
+    @assert !isnothing(state_inits) "state_init must be provided"
+    @assert !isnothing(state_goals) "state_goal must be provided"
+    @assert length(state_data) == length(state_names) == length(state_inits) == length(state_goals) "state_data, state_names, state_inits, and state_goals must have the same length"
 
+    @assert length(control_bounds) == n_control_derivatives + 1 "control_bounds must have $n_control_derivatives + 1 elements"
+
+    # assert that state names are unique
+    @assert length(state_names) == length(Set(state_names)) "state_names must be unique"
+
+    # Control data
+    control_derivative_names = [
+        Symbol("d"^i * string(control_name))
+            for i = 1:n_control_derivatives
+    ]
+
+    control_names = (control_name, control_derivative_names...)
+
+    control_bounds = NamedTuple{control_names}(control_bounds)
+
+    # Timestep data
     if free_time
         if Δt isa Real
             Δt = fill(Δt, 1, T)
@@ -275,94 +306,64 @@ function initialize_unitary_trajectory(
         else
             @assert size(Δt) == (1, T) "Δt must be a Real, AbstractVector, or 1x$(T) AbstractMatrix"
         end
-        timestep = :Δt
+        timestep = timestep_name
     else
         @assert Δt isa Real "Δt must be a Real if free_time is false"
         timestep = Δt
     end
 
-    Ũ⃗_init = operator_to_iso_vec(U_init)
-    if U_goal isa EmbeddedOperator
-        Ũ⃗_goal = operator_to_iso_vec(U_goal.operator)
-    else
-        Ũ⃗_goal = operator_to_iso_vec(U_goal)
-    end
 
     # Constraints
-    Ũ⃗_inits = repeat([Ũ⃗_init], length(Ũ⃗_keys))
     initial = (;
-        (Ũ⃗_keys .=> Ũ⃗_inits)...,
-        a = zeros(n_drives),
+        (state_names .=> state_inits)...,
+        control_name => zeros(n_drives),
     )
 
-    final = (
-        a = zeros(n_drives),
+    final = (;
+        control_name => zeros(n_drives),
     )
 
-    Ũ⃗_goals = repeat([Ũ⃗_goal], length(Ũ⃗_keys))
-    goal = (; (Ũ⃗_keys .=> Ũ⃗_goals)...)
+    goal = (; (state_names .=> state_goals)...)
 
     # Bounds
-    bounds = all_a_bounds
+    bounds = control_bounds
 
-    if bound_unitary
-        Ũ⃗_dim = length(Ũ⃗_init)
-        Ũ⃗_bounds = repeat([(-ones(Ũ⃗_dim), ones(Ũ⃗_dim))], length(Ũ⃗_keys))
-        bounds = merge(bounds, (; (Ũ⃗_keys .=> Ũ⃗_bounds)...))
-    end
-
-    # Initial state and control values
-    if isnothing(a_guess)
-        Ũ⃗ = initialize_unitaries(U_init, U_goal, T, geodesic=geodesic)
-        Ũ⃗_values = repeat([Ũ⃗], length(Ũ⃗_keys))
-        a_values = initialize_controls(
-            n_drives,
-            n_derivatives,
-            T,
-            bounds[a_keys[1]],
-            drive_derivative_σ
-        )
-    else
-        @assert size(a_guess, 1) == n_drives "a_guess must have the same number of drives as n_drives"
-        @assert size(a_guess, 2) == T "a_guess must have the same number of timesteps as T"
-        @assert !isnothing(system) "system must be provided if a_guess is provided"
-
-        if Δt isa AbstractMatrix
-            ts = vec(Δt)
-        elseif Δt isa Float64
-            ts = fill(Δt, T)
-        else
-            ts = Δt
-        end
-
-        if system isa AbstractVector
-            @assert length(system) == length(Ũ⃗_keys) "systems must have the same length as Ũ⃗_keys"
-            Ũ⃗_values = map(system) do sys
-                unitary_rollout(Ũ⃗_init, a_guess, ts, sys; integrator=rollout_integrator)
-            end
-        else
-            Ũ⃗ = unitary_rollout(Ũ⃗_init, a_guess, ts, system; integrator=rollout_integrator)
-            Ũ⃗_values = repeat([Ũ⃗], length(Ũ⃗_keys))
-        end
-        Ũ⃗_values = Matrix{Float64}.(Ũ⃗_values)
-        a_values = initialize_controls(a_guess, ts, n_derivatives)
+    # Put unit box bounds on the state if bound_state is true
+    if bound_state
+        state_dim = length(state_inits[1])
+        state_bounds = repeat([(-ones(state_dim), ones(state_dim))], length(state_names))
+        bounds = merge(bounds, (; (state_names .=> state_bounds)...))
     end
 
     # Trajectory
-    keys = [Ũ⃗_keys..., a_keys...]
-    values = [Ũ⃗_values..., a_values...]
+    if isnothing(a_guess)
+        # Randomly sample controls
+        a_values = initialize_control_trajectory(
+            n_drives,
+            n_control_derivatives,
+            T,
+            bounds[control_name],
+            drive_derivative_σ
+        )
+    else
+        # Use provided controls and take derivatives
+        a_values = initialize_control_trajectory(a_guess, Δt, n_control_derivatives)
+    end
+
+    names = [state_names..., control_names...]
+    values = [state_data..., a_values...]
 
     if free_time
-        push!(keys, :Δt)
+        push!(names, timestep_name)
         push!(values, Δt)
-        controls = (a_keys[end], :Δt)
-        bounds = merge(bounds, (Δt = Δt_bounds,))
+        controls = (control_names[end], timestep_name)
+        bounds = merge(bounds, (; timestep_name => Δt_bounds,))
     else
-        controls = (a_keys[end],)
+        controls = (control_names[end],)
     end
 
     return NamedTrajectory(
-        (; (keys .=> values)...);
+        (; (names .=> values)...);
         controls=controls,
         timestep=timestep,
         bounds=bounds,
@@ -373,97 +374,179 @@ function initialize_unitary_trajectory(
     )
 end
 
-function initialize_quantum_state_trajectory(
-    ψ̃_goals::AbstractVector{<:AbstractVector{<:Real}},
-    ψ̃_inits::AbstractVector{<:AbstractVector{<:Real}},
+function initialize_trajectory(
+    U_goal::OperatorType,
     T::Int,
-    Δt::Union{Real, AbstractVector{<:Real}},
-    n_drives::Int,
-    all_a_bounds::NamedTuple{anames, <:Tuple{Vararg{VectorBound}}} where anames;
-    n_derivatives::Int=2,
-    free_time=false,
-    Δt_bounds::ScalarBound=(0.5 * Δt, 1.5 * Δt),
-    drive_derivative_σ::Float64=0.1,
+    Δt::Union{Real, AbstractVecOrMat{<:Real}},
+    args...;
+    state_name::Symbol=:Ũ⃗,
+    state_names::AbstractVector{<:Symbol}=[state_name],
+    U_init::AbstractMatrix{<:Number}=Matrix{ComplexF64}(I(size(U_goal, 1))),
     a_guess::Union{AbstractMatrix{<:Float64}, Nothing}=nothing,
     system::Union{AbstractQuantumSystem, AbstractVector{<:AbstractQuantumSystem}, Nothing}=nothing,
-    global_data::Union{NamedTuple, Nothing}=nothing,
-    rollout_integrator::Function=exp,
-    ψ̃_keys::AbstractVector{<:Symbol}=[Symbol("ψ̃$i") for i = 1:length(ψ̃_goals)],
-    a_keys::AbstractVector{<:Symbol}=[Symbol("d"^i * "a") for i = 0:n_derivatives]
+    rollout_integrator::Function=expv,
+    geodesic=true,
+    kwargs...
 )
-    @assert length(ψ̃_inits) == length(ψ̃_goals) "ψ̃_inits and ψ̃_goals must have the same length"
-    @assert length(ψ̃_keys) == length(ψ̃_goals) "ψ̃_keys and ψ̃_goals must have the same length"
+    Ũ⃗_init = operator_to_iso_vec(U_init)
 
-    if free_time
-        if Δt isa Real
-            Δt = fill(Δt, 1, T)
-        elseif Δt isa AbstractVector
-            Δt = reshape(Δt, 1, :)
+    if U_goal isa EmbeddedOperator
+        Ũ⃗_goal = operator_to_iso_vec(U_goal.operator)
+    else
+        Ũ⃗_goal = operator_to_iso_vec(U_goal)
+    end
+
+    if isnothing(a_guess)
+        # No guess provided, initialize a geodesic and randomly sample controls
+
+        Ũ⃗_traj = initialize_unitary_trajectory(U_init, U_goal, T; geodesic=geodesic)
+
+        state_data = repeat([Ũ⃗_traj], length(state_names))
+    else
+        if system isa AbstractQuantumSystem
+            @assert size(a_guess, 1) == length(system.H_drives) "a_guess must have the same number of drives as n_drives"
+        elseif system isa AbstractVector
+            @assert size(a_guess, 1) == length(system[1].H_drives) "a_guess must have the same number of drives as n_drives"
+        end
+
+        @assert size(a_guess, 2) == T "a_guess must have the same number of timesteps as T"
+        @assert !isnothing(system) "system must be provided if a_guess is provided"
+
+        if Δt isa AbstractMatrix
+            timesteps = vec(Δt)
+        elseif Δt isa Float64
+            timesteps = fill(Δt, T)
         else
-            @assert size(Δt) == (1, T) "Δt must be a Real, AbstractVector, or 1x$(T) AbstractMatrix"
+            timesteps = Δt
+        end
+
+        if system isa AbstractVector
+            @assert length(system) == length(state_names) "systems must have the same length as state_names"
+            state_data = map(system) do sys
+                unitary_rollout(Ũ⃗_init, a_guess, timesteps, sys; integrator=rollout_integrator)
+            end
+        else
+            state = unitary_rollout(Ũ⃗_init, a_guess, timesteps, system; integrator=rollout_integrator)
+            state_data = [state]
+        end
+        state_data = Matrix{Float64}.(state_data)
+    end
+
+    if system isa AbstractVector && length(state_names) != length(system)
+        state_names = [string(state_name) * "_system_$i" for i = 1:length(system)]
+        @warn "length of state_names and number of systems ($(length(system))) are not equal, created state names for each system: " state_names
+    end
+
+    state_inits = repeat([Ũ⃗_init], length(state_names))
+    state_goals = repeat([Ũ⃗_goal], length(state_names))
+
+    return initialize_trajectory(
+        state_data,
+        state_inits,
+        state_goals,
+        state_names,
+        T,
+        Δt,
+        args...;
+        a_guess=a_guess,
+        kwargs...
+    )
+end
+
+function initialize_trajectory(
+    ψ_goals::AbstractVector{<:AbstractVector{ComplexF64}},
+    ψ_inits::AbstractVector{<:AbstractVector{ComplexF64}},
+    T::Int,
+    Δt::Union{Real, AbstractVector{<:Real}},
+    args...;
+    state_name=:ψ̃,
+    state_names::AbstractVector{<:Symbol}=length(ψ_goals) == 1 ?
+        [state_name] :
+        [Symbol(string(state_name) * "$i") for i = 1:length(ψ_goals)],
+    a_guess::Union{AbstractMatrix{<:Float64}, Nothing}=nothing,
+    system::Union{AbstractQuantumSystem, AbstractVector{<:AbstractQuantumSystem}, Nothing}=nothing,
+    rollout_integrator::Function=expv,
+    kwargs...
+)
+    @assert length(ψ_inits) == length(ψ_goals) "ψ_inits and ψ_goals must have the same length"
+    @assert length(state_names) == length(ψ_goals) "state_names and ψ_goals must have the same length"
+
+    ψ̃_goals = ket_to_iso.(ψ_goals)
+    ψ̃_inits = ket_to_iso.(ψ_inits)
+
+    if isnothing(a_guess)
+        state_data = []
+        for (ψ̃_init, ψ̃_goal) ∈ zip(ψ̃_inits, ψ̃_goals)
+            ψ̃_traj = linear_interpolation(ψ̃_init, ψ̃_goal, T)
+            push!(state_data, ψ̃_traj)
+        end
+        if system isa AbstractVector
+            state_data = repeat(state_data, length(system))
+        end
+    else
+        @assert size(a_guess, 1) == n_drives "a_guess must have n_drives = $(n_drives) drives"
+        @assert size(a_guess, 2) == T "a_guess must have T = $(T) timesteps"
+        @assert !isnothing(system) "system must be provided if a_guess is provided"
+
+        if Δt isa AbstractMatrix
+            timesteps = vec(Δt)
+        elseif Δt isa Float64
+            timesteps = fill(Δt, T)
+        else
+            timesteps = Δt
+        end
+
+        if system isa AbstractVector
+            state_data = []
+            for sys ∈ system
+                for ψ̃_init ∈ ψ̃_inits
+                    ψ̃_traj = rollout(ψ̃_init, a_guess, timesteps, sys;
+                        integrator=rollout_integrator
+                    )
+                    push!(state_data, ψ̃_traj)
+                end
+            end
+        else
+            state_data = []
+            for ψ̃_init ∈ ψ̃_inits
+                ψ̃_traj = rollout(ψ̃_init, a_guess, timesteps, system;
+                    integrator=rollout_integrator
+                )
+                push!(state_data, ψ̃_traj)
+            end
         end
     end
 
-    # Constraints
-    state_initial = (; (ψ̃_keys .=> ψ̃_inits)...)
-    control_initial = (a = zeros(n_drives),)
-    initial = merge(state_initial, control_initial)
+    state_data = Matrix{Float64}.(state_data)
 
-    final = (a = zeros(n_drives),)
-
-    goal = (; (ψ̃_keys .=> ψ̃_goals)...)
-
-    # Bounds
-    bounds = all_a_bounds
-
-    # Initial state and control values
-    if isnothing(a_guess)
-        ψ̃_values = NamedTuple([
-            k => linear_interpolation(ψ̃_init, ψ̃_goal, T)
-                for (k, ψ̃_init, ψ̃_goal) in zip(ψ̃_keys, ψ̃_inits, ψ̃_goals)
-        ])
-        a_values = initialize_controls(
-            n_drives,
-            n_derivatives,
-            T,
-            bounds[a_keys[1]],
-            drive_derivative_σ
-        )
-    else
-        ψ̃_values = NamedTuple([
-            k => rollout(ψ̃_init, a_guess, Δt, system, integrator=rollout_integrator)
-                for (k, ψ̃_init) in zip(ψ̃_keys, ψ̃_inits)
-        ])
-        a_values = initialize_controls(a_guess, Δt, n_derivatives)
+    if system isa AbstractVector
+        if lenth(state_names) != length(system) * length(ψ_goals)
+            state_names = vcat([
+                Symbol.(string.(state_names) .* "_system_$i")
+                    for i = 1:length(system)
+            ]...)
+            @warn "length of state_names and number of systems ($(length(system))) * number of states ($(length(ψ_goals))) are not equal, created state names for each system: " state_names
+        end
+        ψ̃_inits = repeat(ψ̃_inits, length(system))
+        ψ̃_goals = repeat(ψ̃_goals, length(system))
     end
 
-    # Trajectory
-    keys = [ψ̃_keys..., a_keys...]
-    values = [ψ̃_values..., a_values...]
+    state_inits = ψ̃_inits
+    state_goals = ψ̃_goals
 
-    if free_time
-        push!(keys, :Δt)
-        push!(values, Δt)
-        controls = (a_keys[end], :Δt)
-        timestep = :Δt
-        bounds = merge(bounds, (Δt = Δt_bounds,))
-    else
-        controls = (a_keys[end],)
-        @assert Δt isa Real "Δt must be a Real if free_time is false"
-        timestep = Δt
-    end
-
-    return NamedTrajectory(
-        (; (keys .=> values)...);
-        controls=controls,
-        timestep=timestep,
-        bounds=bounds,
-        initial=initial,
-        final=final,
-        goal=goal,
-        global_data=isnothing(global_data) ? (;) : global_data
+    return initialize_trajectory(
+        state_data,
+        state_inits,
+        state_goals,
+        state_names,
+        T,
+        Δt,
+        args...;
+        a_guess=a_guess,
+        kwargs...
     )
 end
+
 
 # ============================================================================= #
 
@@ -536,7 +619,7 @@ end
     drive_bounds = [1.0, 2.0]
     drive_derivative_σ = 0.01
 
-    a, da, dda = TrajectoryInitialization.initialize_controls(n_drives, n_derivates, T, drive_bounds, drive_derivative_σ)
+    a, da, dda = TrajectoryInitialization.initialize_control_trajectory(n_drives, n_derivates, T, drive_bounds, drive_derivative_σ)
 
     @test size(a) == (n_drives, T)
     @test size(da) == (n_drives, T)
@@ -625,10 +708,10 @@ end
     T = 10
     Δt = 0.1
     n_drives = 2
-    all_a_bounds = (a = [1.0, 1.0],)
+    a_bounds = ([1.0, 1.0],)
 
-    traj = initialize_unitary_trajectory(
-        U_goal, T, Δt, n_drives, all_a_bounds
+    traj = initialize_trajectory(
+        U_goal, T, Δt, n_drives, a_bounds
     )
 
     @test traj isa NamedTrajectory
@@ -637,16 +720,16 @@ end
 @testitem "quantum state trajectory initialization" begin
     using NamedTrajectories
 
-    ψ̃_init = ket_to_iso([0.0, 1.0])
-    ψ̃_goal = ket_to_iso([1.0, 0.0])
+    ψ_init = Vector{ComplexF64}([0.0, 1.0])
+    ψ_goal = Vector{ComplexF64}([1.0, 0.0])
 
     T = 10
     Δt = 0.1
     n_drives = 2
-    all_a_bounds = (a = [1.0, 1.0],)
+    all_a_bounds = ([1.0, 1.0],)
 
-    traj = initialize_quantum_state_trajectory(
-        [ψ̃_goal], [ψ̃_init], T, Δt, n_drives, all_a_bounds
+    traj = initialize_trajectory(
+        [ψ_goal], [ψ_init], T, Δt, n_drives, all_a_bounds
     )
 
     @test traj isa NamedTrajectory