+++ docs/src/release_notes.md
@@ -2,11 +2,29 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.3.1] - 2024-10-17
+
+### Fixed
+
+- Fixed and added tests to [`RydbergChainSystem`](@ref)
+
+## [v0.3.0] - 2024-10-10
+
+### Added
+
+- [`PiccoloOptions`] (@ref) to handle custome problem settings.
+
+## Changed
+
+- Refactored trajectory initialization functions
+- Improved documentation
+- Typo fixes
+
 ## [v0.2.0] - 2024-02-22
 
 ### Added
 
-- `EmbeddedOperators` to handle subspace gate optimization and leakage suppression
+- [`EmbeddedOperator`](@ref) to handle subspace gate optimization and leakage suppression
 - Plotting methods for unitary populations
 
 ### Changed